improve: killPort typecheck

+++ packages/@roots/bud-support/src/util/killPort.ts
@@ -1,6 +1,10 @@
 import {exec} from 'child_process'
 
-export const killPort = port =>
+export const killPort = (port: number) => {
+  if (typeof port !== 'number') {
+    throw new Error(`killPort: port must be a number`)
+  }
+
   exec(`kill $(lsof -ti:${port})`, (err, stdout, stderr) => {
     const errors = [stderr, err]
 
@@ -13,3 +17,4 @@ export const killPort = port =>
 
     process.stdout.write(stdout)
   })
+}
